Fix not being able to click on svg elements (by hacking compiled javascript)
Previously attempting to click on an svg element would give the error:

Message: isElementClickable execution failed;
 undefined is not an object (evaluating 'b.length')

This was because in the `IS_ELEMENT_CLICKABLE` atom the javascript code
assumes that the svg element has the `outerHTML` property but it
doesn't. This information is only used for improved error reporting so
it can be safely ignored for svg elements.

[Modern browsers](https://bugs.chromium.org/p/chromium/issues/detail?id=311080)
support `outerHTML` on svg elements so it's unlikely that this will be
fixed upstream.